// libs/core/src/lib/enums/audit-action.enum.ts
export enum AuditAction {
  Create = 'create',
  Update = 'update',
  Delete = 'delete',
  Login = 'login',
  Logout = 'logout',
  Export = 'export',
  ChangeEmail = 'change_email',
  SetPassword = 'set_password',
  Approve = 'approve',
  Reject = 'reject',
  Activate = 'activate',
  Deactivate = 'deactivate',
  ResendInvite = 'resend_invite',
  TwoFactorSetup = 'two_factor_setup',
  TwoFactorEnable = 'two_factor_enable',
  TwoFactorDisable = 'two_factor_disable',
  TwoFactorBackupUsed = 'two_factor_backup_used',
  TwoFactorBackupRegenerate = 'two_factor_backup_regenerate',
  TwoFactorAdminDisable = 'two_factor_admin_disable',
  TwoFactorAdminReset = 'two_factor_admin_reset',
  Suspend = 'suspend',
  Unsuspend = 'unsuspend',
  Reactivate = 'reactivate',
  Restore = 'restore',
  SessionTerminate = 'session_terminate',
  ProviderSwitch = 'provider_switch',
  UnReject = 'unreject',
  TwoFactorEnforcementChange = 'two_factor_enforcement_change',
  UpdateGlobalSettings = 'update_global_settings',
  LogoutAll = 'logout_all',
  BulkAction = 'bulk_action',
  EmailMarkedNotSpam = 'email_marked_not_spam',
  EmailMarkedSpam = 'email_marked_spam',
  EmailArchived = 'email_archived',
  EmailUnarchived = 'email_unarchived',
  EmailTrashed = 'email_trashed',
  EmailRestored = 'email_restored',
  EmailLinkedToCase = 'email_linked_to_case',
  EmailDraftCreated = 'email_draft_created',
  EmailDraftDiscarded = 'email_draft_discarded',
  EmailSent = 'email_sent',
  EmailReplied = 'email_replied',
  EmailForwarded = 'email_forwarded',
  EmailUnlinkedFromCase = 'email_unlinked_from_case',
  EmailLinkedToAgencyRequest = 'email_linked_to_agency_request',
  EmailUnlinkedFromAgencyRequest = 'email_unlinked_from_agency_request',
  EmailMarkedIrrelevant = 'email_marked_irrelevant',
  EmailUnmarkedIrrelevant = 'email_unmarked_irrelevant',
  EmailMarkedRead = 'email_marked_read',
  EmailMarkedUnread = 'email_marked_unread',
  EmailMarkedImportant = 'email_marked_important',
  EmailUnmarkedImportant = 'email_unmarked_important',
  EmailStatusChanged = 'email_status_changed',
  EmailShareLinkRevoked = 'email_share_link_revoked',
  EmailLabelAdded = 'email_label_added',
  EmailLabelRemoved = 'email_label_removed',
  TemplateAutoPopulated = 'template_auto_populated',
  TemplateEmailPreviewed = 'template_email_previewed',
  AgencyDeleteModalOpened = 'agency_delete_modal_opened',
  AgencyMergeModalOpened = 'agency_merge_modal_opened',
  RoleSync = 'role_sync',
  ProfileImageChange = 'profile_image_change',
  AgencyDeleted = 'agency_deleted',
  UndoRestore = 'undo_restore',
}